# ClientGenerator/Models/WeatherForecastCreate.cs
using System;

namespace ClientGenerator.Models
{
    public class WeatherForecastCreate
    {
        public Guid Id { get; set; }

        public DateTime Date { get; set; }

        public int TemperatureC { get; set; }

        public string Summary { get; set; }
    }
}